refactor: asset time stamp and set doc name and result handling

+++ src/libs/satellite/src/db/store.rs
@@ -140,7 +140,7 @@ pub fn set_doc_store(
 ) -> Result<DocContext<DocUpsert>, String> {
     let controllers: Controllers = get_controllers();
 
-    let data = secure_insert_doc(caller, &controllers, collection.clone(), key.clone(), value)?;
+    let data = secure_set_doc(caller, &controllers, collection.clone(), key.clone(), value)?;
 
     Ok(DocContext {
         key,
@@ -149,7 +149,7 @@ pub fn set_doc_store(
     })
 }
 
-fn secure_insert_doc(
+fn secure_set_doc(
     caller: Principal,
     controllers: &Controllers,
     collection: CollectionKey,
@@ -157,10 +157,10 @@ fn secure_insert_doc(
     value: SetDoc,
 ) -> Result<DocUpsert, String> {
     let rule = get_state_rule(&collection)?;
-    insert_doc_impl(caller, controllers, collection, key, value, &rule)
+    set_doc_impl(caller, controllers, collection, key, value, &rule)
 }
 
-fn insert_doc_impl(
+fn set_doc_impl(
     caller: Principal,
     controllers: &Controllers,
     collection: CollectionKey,
@@ -170,18 +170,17 @@ fn insert_doc_impl(
 ) -> Result<DocUpsert, String> {
     let current_doc = get_state_doc(&collection, &key, rule)?;
 
-    match assert_write_permission(
+    assert_write_permission(
         caller,
         controllers,
         &current_doc,
         &rule.write,
+    )?;
+
+    assert_write_timestamp(
+        &current_doc,
         value.updated_at,
-    ) {
-        Ok(_) => (),
-        Err(e) => {
-            return Err(e);
-        }
-    }
+    )?;
 
     assert_description_length(&value.description)?;
 
@@ -324,18 +323,17 @@ fn delete_doc_impl(
 ) -> Result<Option<Doc>, String> {
     let current_doc = get_state_doc(&collection, &key, rule)?;
 
-    match assert_write_permission(
+    assert_write_permission(
         caller,
         controllers,
         &current_doc,
         &rule.write,
+    )?;
+
+    assert_write_timestamp(
+        &current_doc,
         value.updated_at,
-    ) {
-        Ok(_) => (),
-        Err(e) => {
-            return Err(e);
-        }
-    }
+    )?;
 
     delete_state_doc(&collection, &key, rule)
 }
@@ -345,7 +343,6 @@ fn assert_write_permission(
     controllers: &Controllers,
     current_doc: &Option<Doc>,
     rule: &Permission,
-    user_timestamp: Option<u64>,
 ) -> Result<(), String> {
     // For existing collection and document, check user editing is the caller
     if !public_permission(rule) {
@@ -363,6 +360,13 @@ fn assert_write_permission(
         }
     }
 
+    Ok(())
+}
+
+fn assert_write_timestamp(
+    current_doc: &Option<Doc>,
+    user_timestamp: Option<u64>,
+) -> Result<(), String> {
     // Validate timestamp
     match current_doc.clone() {
         None => (),
